feat : 메뉴 display 처리, ApllicationEvent로 처리 추가
