+++ src/main/java/kitchenpos/menus/application/MenuDisplayConditionListener.java
@@ -1,33 +1,28 @@
-package kitchenpos.products.domain;
+package kitchenpos.menus.application;
 
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.UUID;
 
-import kitchenpos.common.DomainService;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.event.TransactionPhase;
+import org.springframework.transaction.event.TransactionalEventListener;
+
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.domain.tobe.domain.ToBeProduct;
-import kitchenpos.products.domain.tobe.domain.ToBeProductRepository;
 
-@DomainService
-public class PriceCompareWithMenuProductService {
-    private final ToBeProductRepository toBeProductRepository;
+@Component
+public class MenuDisplayConditionListener {
+
     private final MenuRepository menuRepository;
 
-    public PriceCompareWithMenuProductService(ToBeProductRepository toBeProductRepository,
-        MenuRepository menuRepository) {
-        this.toBeProductRepository = toBeProductRepository;
+    public MenuDisplayConditionListener(MenuRepository menuRepository) {
         this.menuRepository = menuRepository;
     }
 
-    public void changeMenuDisplay(final UUID productId, final BigDecimal productPrice) {
-        final ToBeProduct toBeProduct = toBeProductRepository.findById(productId)
-            .orElseThrow(NoSuchElementException::new);
-        toBeProduct.changePrice(productPrice);
-
+    @TransactionalEventListener(phase = TransactionPhase.BEFORE_COMMIT)
+    void changeDisplayCondition(UUID productId) {
         final List<Menu> menus = menuRepository.findAllByProductId(productId);
         for (final Menu menu : menus) {
             BigDecimal sum = BigDecimal.ZERO;
